Began step 1 of the algorithm.
Did an initial write up in code of the first step of the algorithm. Beware: it has not undergone testing, and is not yet suitable for use. I was mostly just getting the algorithm down on paper.

I also set up a findings LaTeX file so that we can begin our writeup.