64 node run output (did not finish all)
I also included the remainder of the 32 run, which completed in a
separate job.

The remainder (well, at least the smaller runs so it doesn't spend time
on a 40 minute run at 256 iterations on a large file for a 1h job) of
the 64 node run is in the queue right now.